Player's Melee Combat: Logic Implementation
I've created the animations needed for the Player's NUE to attack uses punches.

SPACEBAR attacks.

The player can kill both fish in Haley's A* Scene. I've still got to create a death and damage animation for the big fish, but I'll do that later.

I'll be working on the fish damage logic tomorrow.