Combat Mechanics: Player Damaged
I've added the ability for players to be damaged on making contact with fish enemies.
Make sure debug mode is on to see the current health of the player. The debug console will output what enemy hit you as well.

Note: I'm still working on the boundary for it so the player is able to safely use weapons, so be aware that it may be difficult to hit things without being hurt at the moment.